refactor: folder structure for send-proof page
