+++ apps/web/src/modules/job/credit/components/SendResultModalBody/styled.ts
