+++ apps/web/src/modules/job/credit/components/SendResultModalBody/types.ts
