+++ apps/web/src/modules/job/credit/pages/send-proof/hooks/useSendingDetails/types.ts
