+++ apps/web/src/modules/job/credit/pages/send-proof/styled.ts
